requirements.txt
aiohttp==0.19.0
beautifulsoup4==4.4.1
chardet==2.3.0
discord.py==0.10.0a0
requests==2.9.1
six==1.10.0
websockets==2.7
wikipedia==1.4.0
wolframalpha==2.2
